requirements.txt
GitPython==3.1.40
neo4j==5.24.0
sqlalchemy==2.0.35
python-dotenv==1.0.0